% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genebabel.R
\docType{package}
\name{genebabel}
\alias{genebabel}
\alias{genebabel-package}
\title{genebabel: A package for converting gene IDs and symbols between different
databases}
\description{
genebabel: A package for converting gene IDs and symbols between different
databases
}
