% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{iterative_select}
\alias{iterative_select}
\title{Iteratively query a database for matches to a query vector.}
\usage{
iterative_select(query, database, match_cols, return_all = FALSE)
}
\arguments{
\item{query}{A character vector.}

\item{database}{A data.frame or tibble  A database to be queried.
See [databases()] for a list of included databases.}

\item{match_cols}{A character vector. The columns in the data to look for matches
with the query. In order of preference, if matches to a column are
found, matches to subsequent columns are not reported by default, unless
return_all is `TRUE`.}

\item{return_all}{A logical indicating whether matches to subsequent columns,
after a match has already been found, should also be returned.}
}
\value{
A tibble containing all rows of the dataframe that matched a query.
  The new column `match__` contains the name of the column that matched
  the query for this row.
}
\description{
`iterative_select()` returns a tibble with all entries of the database that
match the query vector in any of the selected columns.
}
\examples{
iterative_select(c("FLG", "SGK2"),
                 hgnc,
                 c("symbol", "alias_symbol", "prev_symbol"))

}
