% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{join_hgnc}
\alias{join_hgnc}
\title{Join query dataframe with results from  the HUGO Gene Naming Consortium
approved symbol database}
\usage{
join_hgnc(df, query_col, match_cols, select_cols = NULL)
}
\arguments{
\item{df}{A data.frame or tibble. The data used for querying the database.}

\item{query_col}{A character vector of length one. Name of the column in `df`
that will be used to query the database.}

\item{match_cols}{A character vector. The columns in the data to look for matches
with the query. In order of preference, if matches to a column are
found, matches to subsequent columns are not reported by default, unless
return_all is `TRUE`.}

\item{select_cols}{A character vector of column names in the database that
will be merged in the ouput.}
}
\value{
The input dataframe merged with the selected matching columns from
  the HUGO database.
}
\description{
Find which columns are available by checking `help(hgnc)`.
}
\examples{
d <- data.frame(a = 1:3, b = c("FLG", "SGK2", "CDK1"))
join_hgnc(d, "b",
  match_cols = c("symbol", "alias_symbol", "prev_symbol"),
  select_cols = c("entrez_id", "symbol", "refseq_accession"))

}
