% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{join_results}
\alias{join_results}
\title{Join results from a database into an existing dataframe.}
\usage{
join_results(df, query_col, database, match_cols, select_cols = NULL)
}
\arguments{
\item{df}{A data.frame or tibble. The data used for querying the database.}

\item{query_col}{A character vector of length one. Name of the column in `df`
that will be used to query the database.}

\item{database}{A data.frame or tibble  A database to be queried.
See [databases()] for a list of included databases.}

\item{match_cols}{A character vector. The columns in the data to look for matches
with the query. In order of preference, if matches to a column are
found, matches to subsequent columns are not reported by default, unless
return_all is `TRUE`.}

\item{select_cols}{A character vector of column names in the database that
will be merged in the ouput.}
}
\value{
The input dataframe merged with the selected matching columns from
  the database.
}
\description{
`join_results()` queries a database for matches to a column in the supplied
dataset and returns it together with the matches found.
}
\examples{
d <- data.frame(a = 1:3, b = c("FLG", "SGK2", "CDK1"))
join_results(d, "b", hgnc,
  match_cols = c("symbol", "alias_symbol", "prev_symbol"),
  select_cols = c("entrez_id", "symbol", "refseq_accession"))

}
