% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{iterative_select}
\alias{iterative_select}
\title{Iteratively query a database for matches to a query vector.}
\usage{
iterative_select(query, columns, data, return_all = FALSE)
}
\arguments{
\item{query}{A character vector.}

\item{columns}{A character vector. The columns in the data to look for matches
with the query. In order of preference, if matches to a column are
found, matches to subsequent columns are not reported by default, unless
return_all is `TRUE`.}

\item{data}{A data.frame or tibble.}

\item{return_all}{A logical indicating whether matches to subsequent columns,
after a match has already been found, should also be returned. FIXME: Not
implemented!}
}
\value{
A tibble containing all rows of the dataframe that matched a query.
  The new column `match__` contains the name of the column that matched
  the query for this row.
}
\description{
`iterative_join()` returns a tibble with all entries of the database that
match the query vector in any of the selected columns.
}
